JessicaH- merged dayUtil into dateUtil and added tests for calculateDays
