+++ frontend/src/main/utils/dateUtils.js
@@ -6,4 +6,13 @@ const timestampToDate = (timestamp) => {
     return (date.getFullYear() + "-" + (padWithZero(date.getMonth()+1)) + "-" + padWithZero(date.getDate()));
 }
 
-export {timestampToDate, padWithZero};
+
+const calculateDays = (startingDate,currentDate) => {
+    const start = new Date(startingDate);
+    const today = new Date(currentDate);
+    const timeDifference = today.getTime() - start.getTime();
+    const days = Math.floor(timeDifference / (1000 * 3600 * 24));
+    return days;
+  }
+
+export {timestampToDate, padWithZero, calculateDays};
